AddingDummyServicesControllersAndTheirInterfaces

+++ Procuratio.Modules.Cashes.DataAccess/EF/Seeds/ProductionEnviroment/CashStateSeed.cs
@@ -10,11 +10,11 @@ namespace Procuratio.Modules.Cashes.DataAccess.EF.Seeds.ProductionEnviroment
         {
             if (cashStateDbSet.Any()) return;
 
-            cashStateDbSet.Add(new CashState() { ID = (int)CashState.State.Active, StateName = "Activa" });
+            cashStateDbSet.Add(new CashState() { ID = (short)CashState.State.Active, StateName = "Activa" });
 
-            cashStateDbSet.Add(new CashState() { ID = (int)CashState.State.Inactive, StateName = "Inactiva" });
+            cashStateDbSet.Add(new CashState() { ID = (short)CashState.State.Inactive, StateName = "Inactiva" });
 
-            cashStateDbSet.Add(new CashState() { ID = (int)CashState.State.CashClosed, StateName = "Caja cerrada" });
+            cashStateDbSet.Add(new CashState() { ID = (short)CashState.State.CashClosed, StateName = "Caja cerrada" });
         }
     }
 }
