+++ Procuratio.Modules.Cashes.Domain/Entities/Cash.cs
@@ -19,7 +19,7 @@ namespace Procuratio.Modules.Cashes.Domain.Entities
         public int MountTypeID { get; set; }
         public MountType MountType { get; set; }
 
-        public int CashStateID { get; internal set; }
+        public short CashStateID { get; internal set; }
         public CashState CashState { get; set; }
 
         public int UserID { get; set; }
