+++ Procuratio.Modules.Menues.DataAccess/EF/Seeds/ProductionEnviroment/CategoryItemStateSeed.cs
@@ -10,9 +10,9 @@ namespace Procuratio.Modules.Menues.DataAccess.EF.Seeds.ProductionEnviroment
         {
             if (categoryItemStateDbSet.Any()) return;
 
-            categoryItemStateDbSet.Add(new CategoryItemState() { ID = (int)CategoryItemState.State.Available, StateName = "Disponible" });
+            categoryItemStateDbSet.Add(new CategoryItemState() { ID = (short)CategoryItemState.State.Available, StateName = "Disponible" });
 
-            categoryItemStateDbSet.Add(new CategoryItemState() { ID = (int)CategoryItemState.State.Deleted, StateName = "Eliminado" });
+            categoryItemStateDbSet.Add(new CategoryItemState() { ID = (short)CategoryItemState.State.Deleted, StateName = "Eliminado" });
         }
     }
 }
