+++ Procuratio.Modules.Orders.DataAccess/EF/Seeds/ProductionEnviroment/TakeAwayStateSeed.cs
@@ -10,11 +10,11 @@ namespace Procuratio.Modules.Orders.DataAccess.EF.Seeds.ProductionEnviroment
         {
             if (takeAwayStateDbSet.Any()) return;
 
-            takeAwayStateDbSet.Add(new TakeAwayState() { ID = (int)TakeAwayState.State.InProgress, StateName = "En curso" });
+            takeAwayStateDbSet.Add(new TakeAwayState() { ID = (short)TakeAwayState.State.InProgress, StateName = "En curso" });
 
-            takeAwayStateDbSet.Add(new TakeAwayState() { ID = (int)TakeAwayState.State.Completed, StateName = "Completado" });
+            takeAwayStateDbSet.Add(new TakeAwayState() { ID = (short)TakeAwayState.State.Completed, StateName = "Completado" });
 
-            takeAwayStateDbSet.Add(new TakeAwayState() { ID = (int)TakeAwayState.State.DidNotCome, StateName = "No vino" });
+            takeAwayStateDbSet.Add(new TakeAwayState() { ID = (short)TakeAwayState.State.DidNotCome, StateName = "No vino" });
         }
     }
 }
