+++ Procuratio.Modules.Orders.Domain/Entities/Reserve.cs
@@ -17,7 +17,7 @@ namespace Procuratio.Modules.Orders.Domain.Entities
         public int? OrderID { get; set; }
         public Order Order { get; set; }
 
-        public int ReserveStateID { get; internal set; }
+        public short ReserveStateID { get; internal set; }
         public ReserveState ReserveState { get; set; }
 
         public int CustomerID { get; set; }
